Initial commit: Adding the script to retrieve shortcut files and its information
Written this following the #Asrmageddon on January 13th 2013. 

The script retrieves shortcut files and their information, such as name, path, target path, arguments, start in and description, and exports them to a CSV file. It also checks for empty subfolders that could indicate something missing. This script can be a valuable tool for system administrators managing large numbers of machines under the context of pulling this information from a non-affected machine or gold build or using the missing element to identify what has been affected without using KQL or non-OS-native tooling.